Adding simpler debugging to Parsible.
It is now possible to debug your processors as your run them through
Parsible.  When using '--debug True' the processors loop will no longer
swallow all exceptions and will instead print out a traceback to stdout.

In the future this will be expanded to allow individual processors to
import the debugging logger so that more fine grainer errors can be
caught for further analysis.

By default debugging is disabled and no changes to any scripts need to
be made to update Parsible.