Added batch mode for processing static files.
Parsible will now accept the --batch flag (-b) which changes logic
used to follow the log file.  Parsible will start at the beginning
of the specified file instead of initially seeking to the end.
Furthermore, when Parsible reaches the end of the log file it will
exit and clean up after itself.